+++ react_on_rails_pro/spec/dummy/client/app/startup/SetTimeoutClientApp.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default () => (
+  <div>Disable javascript in your browser options to confirm this value is set by setTimeout during SSR.</div>
+);