+++ react_on_rails_pro/spec/dummy/config/initializers/react_on_rails_pro.rb
@@ -7,8 +7,6 @@ ReactOnRailsPro.configure do |config|
 
   config.rendering_returns_promises = true
 
-  config.include_execjs_polyfills = false
-
   config.server_renderer = "NodeRenderer"
   # If you want Honeybadger or Sentry on the Node renderer side to report rendering errors
   config.throw_js_errors = false